Convert and move should clone test
